append entry - check for log term match

+++ src/raft/vote-request.go
@@ -3,6 +3,6 @@ package raft
 type voteRequest struct {
 	from         peer
 	term         uint64
-	lastLogIndex uint
+	lastLogIndex uint64
 	lastLogTerm  uint64
 }
